Setup a font to be used in Shiny or Markdown

Setup a font to be used in Shiny or Markdown

setup_font(id, output_dir, variants = NULL, prefer_local_source = TRUE)

Arguments

id
Id of the font, correspond to column id from get_all_fonts.

output_dir
Output directory where to save font and CSS files. Must be a directory.

variants
Variant(s) to download, default is to includes all available ones.

prefer_local_source
Generate CSS font-face rules in which user installed fonts are preferred. Use FALSE if you want to force the use of the downloaded font.

Value

Two directories will be created (if they do not exist): fonts and css.

Examples

if (FALSE) {

path_to_www <- tempfile()

# In a Shiny app, you can use the www/ directory
# in Markdown, use a subfolder of your Rmd directory
setup_font(
  id = "open-sans-condensed",
  output_dir = path_to_www
)

unlink(path_to_www)

}